Remove non specific code for windows/nix
